refactor: extract XDR codec functions into separate files
Split NFS v3 handler files to separate XDR encoding/decoding logic:

- Extract _codec.go files for all 22 NFS procedures
- Extract read_cache.go, read_content.go, read_prefetch.go, read_validation.go
- Extract write_validation.go

This improves code organization by separating protocol serialization
from business logic. All handler files are now under 700 lines.

Files created:
- 22 *_codec.go files (4,017 lines total)
- 5 read_*.go helper files
- 1 write_validation.go file

No functional changes - pure structural refactoring.
